requirements.txt
google-cloud-aiplatform
google-cloud-storage
python-dotenv
ffmpeg-python
